requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-dotenv
jinja2
weasyprint
redis
celery
email-validator
httpx
crawl4ai
pytest
PyJWT
passlib[bcrypt]
python-multipart
ruff
